kafka,tools,main: refactor WithSignalCancel and Getenv to tools package
  move the WithSignalCancel and getenv functions which were previously
  in the consumer.go file into a new tools.go file. this seperates the
  responsibility of utility functions to a new tools package while
  keeping the kafka package clean
