+++ backend/internal/kafka/consumer.go
@@ -3,15 +3,13 @@ package kafka
 import (
 	"context"
 	"log/slog"
-	"os"
-	"os/signal"
-	"syscall"
 
+	"github.com/Prypiatos/ems-app/backend/internal/tools"
 	ckafka "github.com/confluentinc/confluent-kafka-go/v2/kafka"
 )
 
 func NewConsumer(topic, groupID string) (*ckafka.Consumer, error) {
-	broker := getenv("KAFKA_BROKER", "localhost:9092")
+	broker := tools.Getenv("KAFKA_BROKER", "localhost:9092")
 
 	c, err := ckafka.NewConsumer(&ckafka.ConfigMap{
 		"bootstrap.servers": broker,
@@ -65,25 +63,4 @@ func Consume(ctx context.Context, c *ckafka.Consumer) {
 			}
 		}
 	}
-}
-
-func WithSignalCancel() context.Context {
-	ctx, cancel := context.WithCancel(context.Background())
-
-	go func() {
-		ch := make(chan os.Signal, 1)
-		signal.Notify(ch, syscall.SIGINT, syscall.SIGTERM)
-		<-ch
-		slog.Info("signal received, cancelling context")
-		cancel()
-	}()
-
-	return ctx
-}
-
-func getenv(k, def string) string {
-	if v := os.Getenv(k); v != "" {
-		return v
-	}
-	return def
 }
